refactor: sync-data-handler converting it to hook
